refactor operation factory
