+++ fearlessTests/Mocks/ModuleMocks.swift
@@ -20627,7 +20627,6 @@ import Cuckoo
 @testable import fearless
 
 import Foundation
-import IrohaCrypto
 import RobinHood
